fixed 4th collector with working authentication

+++ kuhako/src/main/java/com/capstone/kuhako/services/CollectorModuleServices/PaymentAssuranceServiceImpl.java
@@ -1,7 +1,9 @@
 package com.capstone.kuhako.services.CollectorModuleServices;
 
+import com.capstone.kuhako.models.Collector;
 import com.capstone.kuhako.models.CollectorModules.PaymentAssurance;
 import com.capstone.kuhako.repositories.CollectorModuleRepository.PaymentAssuranceRepository;
+import com.capstone.kuhako.repositories.CollectorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,27 +13,46 @@ import org.springframework.stereotype.Service;
 public class PaymentAssuranceServiceImpl implements PaymentAssuranceService {
     @Autowired
     private PaymentAssuranceRepository paymentAssuranceRepository;
+    @Autowired
+    private CollectorRepository collectorRepository;
 
-    public void createPaymentAssurance(PaymentAssurance paymentAssurance){
-//      User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(stringToken));
-        PaymentAssurance newPaymentAssurance = new PaymentAssurance();
-        newPaymentAssurance.setCollector(paymentAssurance.getCollector());
-        newPaymentAssurance.setPaymentStatus(paymentAssurance.getPaymentStatus());
-        paymentAssuranceRepository.save(newPaymentAssurance);
+    public void createPaymentAssurance(Long collectorId, PaymentAssurance paymentAssurance){
+        Collector collector = collectorRepository.findById(collectorId).orElse(null);
+        if (collector != null) {
+            paymentAssurance.setCollector(collector);
+            paymentAssuranceRepository.save(paymentAssurance);
+        }
     }
     public Iterable<PaymentAssurance> getPaymentAssurance(){
         return paymentAssuranceRepository.findAll();
     }
 
-    public ResponseEntity deletePaymentAssurance(Long id){
-        paymentAssuranceRepository.deleteById(id);
-        return new ResponseEntity<>("PaymentAssurance Deleted successfully", HttpStatus.OK);
+    public Iterable<PaymentAssurance> getPaymentAssuranceByCollectorId(Long collectorId) {
+        return paymentAssuranceRepository.findPaymentAssuranceByCollectorId(collectorId);
     }
-    public ResponseEntity updatePaymentAssurance(Long id, PaymentAssurance paymentAssurance){
-        PaymentAssurance paymentAssuranceForUpdate = paymentAssuranceRepository.findById(id).get();
-        paymentAssuranceForUpdate.setCollector(paymentAssurance.getCollector());
-        paymentAssuranceForUpdate.setPaymentStatus(paymentAssurance.getPaymentStatus());
-        paymentAssuranceRepository.save(paymentAssuranceForUpdate);
-        return new ResponseEntity("PaymentAssurance updated successfully", HttpStatus.OK);
+
+    public ResponseEntity deletePaymentAssurance(Long collectorId, Long id){
+        PaymentAssurance paymentAssuranceToDelete = paymentAssuranceRepository.findById(id).orElse(null);
+
+        if (paymentAssuranceToDelete != null && paymentAssuranceToDelete.getCollector().getCollector_id().equals(collectorId)) {
+            paymentAssuranceRepository.deleteById(id);
+            return new ResponseEntity<>("PaymentAssurance Deleted successfully", HttpStatus.OK);
+        } else if (paymentAssuranceToDelete != null) {
+            return new ResponseEntity<>("PaymentAssurance does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("PaymentAssurance not found", HttpStatus.NOT_FOUND);
+        }
+    }
+    public ResponseEntity updatePaymentAssurance(Long collectorId, Long id, PaymentAssurance paymentAssurance){
+        PaymentAssurance paymentAssuranceForUpdate = paymentAssuranceRepository.findById(id).orElse(null);
+        if (paymentAssuranceForUpdate != null && paymentAssuranceForUpdate.getCollector().getCollector_id().equals(collectorId)) {
+            paymentAssuranceForUpdate.setPaymentStatus(paymentAssurance.getPaymentStatus());
+            paymentAssuranceRepository.save(paymentAssuranceForUpdate);
+            return new ResponseEntity<>("PaymentAssurance Updated successfully", HttpStatus.OK);
+        } else if (paymentAssuranceForUpdate != null) {
+            return new ResponseEntity<>("PaymentAssurance does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("PaymentAssurance not found",HttpStatus.NOT_FOUND);
+        }
     }
 }
